+++ data/src/main/kotlin/com/wepli/data/applemusic/common/response/AppleAlbumResponse.kt
@@ -1,6 +1,7 @@
 package com.wepli.data.applemusic.common.response
 
 import com.wepli.data.applemusic.common.response.base.AppleArtworkResponse
+import com.wepli.data.applemusic.common.response.base.AppleRelationshipsResponse
 import kotlinx.serialization.Serializable
 import model.album.Album
